+++ CHANGELOG.md
@@ -11,6 +11,10 @@
 
 * Fixed a rare race in `Conn.start` that could cause goroutines to be leaked if the provided context was canceld/expired.
 
+### Other Changes
+
+* The field `MessageHeader.Durable` is not omitted when it's `false`.
+
 ## 1.1.0 (2024-08-20)
 
 ### Features Added